chore: review and clean up code structure across multiple files

+++ src/cpp/pricing.h
@@ -5,13 +5,13 @@
 #include <vector>
 #include <unordered_map>
 
-// Calculate the final selling price of a product
-double calculateFinalPrice(
+// Calculate the final selling price of a product (in cents)
+int calculateFinalPrice(
     const std::string &productName,
     const std::vector<std::string> &currentEffects,
     const std::unordered_map<std::string, int> &effectMultipliers);
 
-// Calculate the total cost of a mix
-double calculateFinalCost(
+// Calculate the total cost of a mix (in cents)
+int calculateFinalCost(
     const MixState &mixState,
     const std::vector<Substance> &substances);
